requirements.txt
azure-identity
azure-mgmt-compute
azure-mgmt-resource
Flask
python-dotenv
redis
